added home controller and views, modified productcontroller

+++ EekiBooksOnline/Areas/Admin/Controllers/ProductController.cs
@@ -6,6 +6,141 @@ using EekiBooks.Models.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 
+//namespace EekiBooksOnline.Areas.Admin.Controllers
+//{
+//    public class ProductController : Controller
+//    {
+//        private readonly IUnitOfWork _unitOfWork;
+//        private readonly IWebHostEnvironment _webHostEnvironment;
+
+//        public ProductController(IUnitOfWork unitOfWork, IWebHostEnvironment webHostEnvironment)
+//        {
+//            _unitOfWork = unitOfWork;
+//            _webHostEnvironment = webHostEnvironment;
+//        }
+//        public IActionResult Index()
+//        {
+
+//            return View();
+//        }
+
+//        // GET
+//        public IActionResult Upsert(int? id)
+//        {
+//            ProductVM productVM = new()
+//            {
+//                Product = new(),
+//                CategoryList = _unitOfWork.Category.GetAll().Select(i => new SelectListItem
+//                {
+//                    Text = i.Name,
+//                    Value = i.Id.ToString()
+//                }),
+
+//                CoverTypeList = _unitOfWork.CoverType.GetAll().Select(i => new SelectListItem
+//                {
+//                    Text = i.Name,
+//                    Value = i.Id.ToString()
+//                }),
+
+//            };
+//            if (id == null || id == 0)
+//            {
+//                // create producct
+//                return View(productVM);
+//            }
+//            else
+//            {
+//                // update product
+//                productVM.Product = _unitOfWork.Product.GetFirstOrDefault(u => u.Id == id);
+//                return View(productVM);
+//            }
+//        }
+
+//        // POST
+//        [HttpPost]
+//        [ValidateAntiForgeryToken]
+//        public IActionResult Upsert(ProductVM obj, IFormFile? file)
+//        {
+
+//            if (ModelState.IsValid)
+//            {
+//                string wwRootPath = _webHostEnvironment.WebRootPath;
+//                if (file != null)
+//                {
+//                    string fileName = Guid.NewGuid().ToString();
+//                    var uploads = Path.Combine(wwRootPath, @"images\products");
+//                    var extension = Path.GetExtension(file.FileName);
+
+//                    if (obj.Product.ImageUrl != null)
+//                    {
+//                        var oldImagePath = Path.Combine(wwRootPath, obj.Product.ImageUrl.TrimStart('\\'));
+//                        if(System.IO.File.Exists(oldImagePath))
+//                        {
+//                            System.IO.File.Delete(oldImagePath);
+//                        }
+//                    }
+//                    using (var fileStreams = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
+//                    {
+//                        file.CopyTo(fileStreams);
+//                    }
+//                    obj.Product.ImageUrl = @"\images\products\" + fileName + extension;
+//                }
+//                if (obj.Product.Id == 0)
+//                {
+//                    _unitOfWork.Product.Add(obj.Product);
+//                }
+//                else
+//                {
+//                    _unitOfWork.Product.Update(obj.Product);
+//                }
+//                _unitOfWork.Save();
+//                TempData["Success"] = "Product created successfully";
+//                return RedirectToAction("Index");
+//            }
+//            return View(obj);
+//        }
+
+
+//        #region API CALLS
+//        [HttpGet]
+//        public IActionResult GetAll()
+//        {
+//            var productList = _unitOfWork.Product.GetAll(includeProperties: "Category,CoverType");
+//            return Json(new { data = productList });
+//        }
+
+
+//        [HttpDelete]
+//        public IActionResult Delete(int? id)
+//        {
+//            var obj = _unitOfWork.Product.GetFirstOrDefault(u => u.Id == id);
+//            if (obj == null)
+//            {
+//                return Json(new { success = false, message = "Error while deleting" });
+//            }
+
+//            var oldImagePath = obj.ImageUrl;
+//            if (!string.IsNullOrEmpty(oldImagePath))
+//            {
+//                oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, oldImagePath.TrimStart('\\'));
+//                if (System.IO.File.Exists(oldImagePath))
+//                {
+//                    System.IO.File.Delete(oldImagePath);
+//                }
+//            }
+//            _unitOfWork.Product.Remove(obj);
+//            _unitOfWork.Save();
+
+//            return Json(new { success = true, message = "Book successfully deleted" });
+//        }
+//        #endregion
+
+//    }
+
+//}
+
+
+
 namespace EekiBooksOnline.Areas.Admin.Controllers
 {
     public class ProductController : Controller
@@ -21,16 +156,14 @@ namespace EekiBooksOnline.Areas.Admin.Controllers
 
         public IActionResult Index()
         {
-
             return View();
         }
 
-        // GET 
         public IActionResult Upsert(int? id)
         {
             ProductVM productVM = new()
             {
-                Product = new(),
+                Product = new Product(),
                 CategoryList = _unitOfWork.Category.GetAll().Select(i => new SelectListItem
                 {
                     Text = i.Name,
@@ -44,51 +177,47 @@ namespace EekiBooksOnline.Areas.Admin.Controllers
                 }),
 
             };
-            if (id == null || id == 0)
+            if (id != null && id > 0)
             {
-                // create product
-                //ViewBag.CategoryList = CategoryList;
-                //ViewData["CoverTypeList"] = CoverTypeList;
-                return View(productVM);
+                var existingProduct = _unitOfWork.Product.GetFirstOrDefault(u => u.Id == id);
+                if (existingProduct != null)
+                {
+                    productVM.Product = existingProduct;
+                }
+                else
+                {
+                    return NotFound(); // Handle the case where the product doesn't exist
+                }
             }
-            else
-            {
-                // update product
-                productVM.Product = _unitOfWork.Product.GetFirstOrDefault(u => u.Id == id);
-                return View(productVM);
-            }
-
+            return View(productVM);
         }
 
-        // POST
         [HttpPost]
         [ValidateAntiForgeryToken]
         public IActionResult Upsert(ProductVM obj, IFormFile? file)
         {
-
             if (ModelState.IsValid)
             {
                 string wwRootPath = _webHostEnvironment.WebRootPath;
                 if (file != null)
                 {
                     string fileName = Guid.NewGuid().ToString();
-                    var uploads = Path.Combine(wwRootPath, @"images\products");
+                    var uploads = Path.Combine(wwRootPath, "images/products");
                     var extension = Path.GetExtension(file.FileName);
 
                     if (obj.Product.ImageUrl != null)
                     {
                         var oldImagePath = Path.Combine(wwRootPath, obj.Product.ImageUrl.TrimStart('\\'));
-                        if(System.IO.File.Exists(oldImagePath))
+                        if (System.IO.File.Exists(oldImagePath))
                         {
                             System.IO.File.Delete(oldImagePath);
                         }
                     }
-
                     using (var fileStreams = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
                     {
                         file.CopyTo(fileStreams);
                     }
-                    obj.Product.ImageUrl = @"\images\products\" + fileName + extension;
+                    obj.Product.ImageUrl = "/images/products/" + fileName + extension;
                 }
                 if (obj.Product.Id == 0)
                 {
@@ -101,7 +230,6 @@ namespace EekiBooksOnline.Areas.Admin.Controllers
                 _unitOfWork.Save();
                 TempData["Success"] = "Product created successfully";
                 return RedirectToAction("Index");
-
             }
             return View(obj);
         }
@@ -115,26 +243,28 @@ namespace EekiBooksOnline.Areas.Admin.Controllers
         }
 
         [HttpDelete]
-      
         public IActionResult Delete(int? id)
         {
-      
             var obj = _unitOfWork.Product.GetFirstOrDefault(u => u.Id == id);
             if (obj == null)
             {
-                return Json(new {success = false, message="Error while deleting"});
+                return Json(new { success = false, message = "Error while deleting" });
             }
-            var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, obj.ImageUrl.TrimStart('\\'));
-            if (System.IO.File.Exists(oldImagePath))
+
+            var oldImagePath = obj.ImageUrl;
+            if (!string.IsNullOrEmpty(oldImagePath))
             {
-                System.IO.File.Delete(oldImagePath);
+                oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, oldImagePath.TrimStart('\\'));
+                if (System.IO.File.Exists(oldImagePath))
+                {
+                    System.IO.File.Delete(oldImagePath);
+                }
             }
             _unitOfWork.Product.Remove(obj);
             _unitOfWork.Save();
+
             return Json(new { success = true, message = "Book successfully deleted" });
         }
         #endregion
-
     }
-
 }
